requirements.txt
accelerate==0.18.0
diffusers==0.16.1
Flask==2.3.1
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.15.0
flask-restx==1.1.0
marshmallow-sqlalchemy==0.29.0
numpy==1.24.3
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
Pillow==9.5.0
python-dotenv==1.0.0
safetensors==0.3.1
segment-anything==1.0
simple-websocket==0.10.0
torch==2.0.0+rocm5.4.2
torchaudio==2.0.1+rocm5.4.2
torchvision==0.15.1+rocm5.4.2
transformers==4.28.1